Added in "virtual triggers" as a concept that triggers_and_io now supports.
It saves a map of virtual triggers and their options to settings'
'virtually triggered actions' key. You can observe it to see when it
changes. This will get used to construct an api call to ioMapping to set
all virtual triggers high or low when whatever application logic is driving
it wants a trigger event. Proper soft triggers, writing a bit to the fpga
to fiddle with the trigger system, aren't a thing because that would
involve hardware interrupts on the CPU. So I'm told. 😕