Add asset generation during build.
When started as a module, chronosGui2 no longer has the luxury of being
able to load files relative to the current working directory. Instead we
can compile the graphical assets into a generated python file.

This means that the developer needs to do an 'python3 setup.py build_ui'
before the GUI is effectively able to boot to ensure that the generated
files exist.