Use generated UI files when loading screen classes.
UI files can be built during packaging, or by doing make in the
'generated' directory. We can therefore load the prebuild files
instead of doing it on each boot of the UI.

Not all screens can do this yet, since we have some weirdness with
filenames that include dots or hyphens that break pythons import
handling.