Added libqt5svg5 as an explicit dependency of chronos-gui-2.
This was caused by a bug report where I couldn't get the battery icon to display. It was traced down to SVG images not rendering on my camera. The reason this wasn't working is as follows:

Foobar, [20.01.20 19:04]
Okay, the reason that it works for me is that libqt5svg5 is a package that is 'recommended' by libqt5gui5, but it is not required.

This means that when you do an apt-get install, it will possibly come in as one of the many packages that gets auto-selected by the package manager, but it is not guaranteed to be installed.

Foobar, [20.01.20 19:05]
When I spun up this camera to test, I just grabbed a working camApp camera, and threw an 'apt-get install chronos-gui2' at it, which apparently by default also installed libqt5svg5

Foobar, [20.01.20 19:06]
So, my recomendation is to add libqt5svg5 as an explicit dependency of chronos-gui2 in the debian/control file