Implemented export and import cal data buttons, connected safely remove button.
The export and import cal data buttons can now be used to make API calls
for flat field exports and cal .bin imports to the camera. The safely
remove button in the Saved Video Location screen had not been bound
to an action. It now pulls up the storage screen where the user may
click unmount and receive a visual confirmation of a successful
unmount.

TODO: add error and success messaging to the export/import cal buttons.